requirements.txt
pytest
playwright
pytest-playwright
boto3
python-dotenv
allure-pytest
logzero
jsonschema
pytest-xdist
faker
hypothesis
allure-python-commons
git+ssh://git@github.com/dock-tech/allure-pytest-dock-lib.git@master